pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chronicles"
version = "0.1.0"
description = "Automatic knowledge extraction from agent session transcripts"
requires-python = ">=3.11"
license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Documentation",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "jinja2>=3.1",
    "json-repair>=0.58.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
tfidf = [
    "scikit-learn>=1.4",
]
dev = [
    "pytest>=8.0",
    "pytest-tmp-files>=0.0.2",
    "scikit-learn>=1.4",
]

[project.scripts]
chronicles = "chronicles.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
chronicles = ["templates/*.j2"]

[tool.pytest.ini_options]
testpaths = ["tests"]
